refactor(observability): extract query/ingestion leaves to observability-engine (Slice 2, ~4K)
Slice 2 of the observability decomposition: move 5 foundation-pure leaves
into the proximadb-observability-engine crate — query::{promql,
tantivy_log_index, metrics} and ingestion::{buffer, parser}. ~4K more root
test-binary reduction (those modules' inline #[cfg(test)] suites now compile
into the engine crate's binary, not the root's).

These leaves are foundation-clean (cross-checked against .victor/project.db +
compile-time grep): zero ObservabilityStorage refs, zero facade-type refs,
only proto + self `crate::` deps. Moved with the `crate::observability::` →
`crate::` import rewrite + a new crate-level `proto` re-export module that
mirrors the root `src/proto/mod.rs`.

The observability FACADE remains in root (bidirectionally entangled):
- mod.rs (Service), ingestion/mod.rs (Ingester), query/mod.rs (QueryEngine),
  query/logs — constructed by network handlers; the OTLP adapters wrap
  ObservabilityService; root's query-facade strategy reaches storage via
  query_engine.storage().
- traces + graph_linking couple to TraceSpan (defined in root storage/).
- rank_metrics couples to the rank-onnx modality.
A clean facade extraction needs TWO ports (ObservabilityStoragePort ~4
methods for the QueryEngine accessor + a service-port for the OTLP↔Service
coupling) + TraceSpan relocation — deferred to Slice 3.

Root query/mod.rs + ingestion/mod.rs re-export the moved leaves, so all
callers (crate::observability::query::promql, etc.) are unchanged. Storage
up-edge ratchet unchanged at 197.

Verified: cargo check --lib + --tests (root + crate); crate clippy --lib --bins
(CI form) AND --all-targets -D warnings; fmt; layering.
