+++ crates/modalities/proximadb-observability-engine/src/ingestion/buffer.rs
