+++ crates/modalities/proximadb-observability-engine/src/query/tantivy_log_index.rs
